Added pthread library dependencies for Linux
pthread was used before, but not specified as dependency.
This was working because tools added that dependency anyway.
It is now explicitly added as dependency to some external functions.

"Dummy" pthread libraries are provided for Windows. The Windows code doesn't depend on pthread, but compilation needs to find some libary of this name (see MDDDummyPthreadLibrary.c).