Commented out section "Raw Data Buffer"
I found it difficult to follow this section and it also seems not
to be very essential in the remaining context.

Also, I'm not sure about the claim that "Modelica variables of type
\modelica{String} are immutable". The default variability of String
is"discrete-time" (MLS v3.3, p. 37) and thus it can be changed at
events. I guess the immutability refers to the external function
interface, but I find the text not so clear.