Commented out two figures in the embedded section
The figures are a bit hard to read and I wanted to save some space so
that we don't have just a few lines that start a new page
(this reduces the paper to 11 pages).
The figures are not so essential to understand the text, 
though it could be nice to have them if one
could improve them so that they depict things more readable and compact.
However, time is running out...